refactor: extract shared media module, centralize sender context
- Extract image resize/compress, download+encode, and STT download
  from discord.rs into shared media.rs module
- Both Discord and Slack adapters now use media::download_and_encode_image()
  and media::download_and_transcribe() with optional auth_token parameter
- Move sender context XML wrapping into AdapterRouter.handle_message()
  (was duplicated in each adapter)
- Add image/audio attachment support to Slack adapter via files[] in
  app_mention events (requires files:read bot scope)
- Slack file downloads use Bearer token auth (private URLs)
